refactor(theme): extract theme into /lib
